+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Improve openapi parameters object, dump attribute descriptions if any
 
+* Instantiate example URLs, via the first test_case filling the service
+  parameters.
+
 ## 0.27.1 - 2025-03-22
 
 * Same as 0.27.0 but docker images fixed.